add skip tag here

+++ android/adbPullVideo.sh
@@ -7,29 +7,31 @@ destination_folder_on_host="/media/pi/ssd"
 temp_file_list="videos.txt"
 sanitize_script="./formatFile.py"
 format_flag=false
+skip_pull=0
 
-while getopts "p:d:f" opt; do
+while getopts "p:d:fs" opt; do
   case ${opt} in
     p ) parent_folder="$OPTARG" ;;
     d ) destination_folder_on_host="$OPTARG" ;;
     f ) format_flag=true ;;
+    s ) skip_pull=1 ;;
     * ) 
-      echo "$(timestamp) Usage: $0 [-d destination_folder]"
+      echo "$(timestamp) Usage: $0 [-d destination_folder] [-f] [-s]"
       exit 1;; 
     esac
   done
-  shift $((OPTIND - 1))
+shift $((OPTIND - 1))
 
-  # Function to get the current timestamp
-  timestamp() {
-    date '+%Y-%m-%d %H:%M:%S'
-  }
+# Function to get the current timestamp
+timestamp() {
+  date '+%Y-%m-%d %H:%M:%S'
+}
 
-  echo "$(timestamp) Destination folder set to: $destination_folder_on_host"
-  mkdir -p "$destination_folder_on_host"
+echo "$(timestamp) Destination folder set to: $destination_folder_on_host"
+mkdir -p "$destination_folder_on_host"
 
-  # Clean up any existing temp file
-  > "$temp_file_list"
+# Clean up any existing temp file
+> "$temp_file_list"
 
 # Capture the file list using adb shell and store it in a temporary file
 adb shell <<EOF > "$temp_file_list"
@@ -51,6 +53,13 @@ find "$parent_folder" -type f \( -iname "*.mp4" -o -iname "*.avi" -o -iname "*.m
   fi
 done
 EOF
+
+# Exit early if skip_pull is set
+if [ $skip_pull -eq 1 ]; then
+  echo "$(timestamp) Skipping video pull as --skip is provided."
+  exit 0
+fi
+
 trap 'echo "Interrupted. Cleaning up $sanitized_filename"; [ -f "$sanitized_filename" ] && rm -f "$sanitized_filename"; exit 1' SIGINT
 
   # Now read the file paths and pull each file to the host machine
